count playlist

+++ app/Console/Commands/MakePlaylist.php
@@ -49,28 +49,30 @@ class MakePlaylist extends Command
         $playlist = $this->playlistRepository
             ->getModel()
             ->with("audio")->findOrFail($this->argument('id'));
-        $hlsDir = storage_path('app/public/hls/' . $playlist->folder);
-        if(File::exists($hlsDir)){
-            File::deleteDirectory($hlsDir);
+        if(count($playlist->audio) > 0){
+            $hlsDir = storage_path('app/public/hls/' . $playlist->folder);
+            if(File::exists($hlsDir)){
+                File::deleteDirectory($hlsDir);
+            }
+            File::makeDirectory($hlsDir, 0777, true);
+
+            $cmd = 'ffmpeg ';
+            foreach ($playlist->audio as $audio){
+                $cmd .= ' -i ' . storage_path('app/'.$audio->path);
+            }
+            $cmd .= ' -filter_complex \'[0:0][1:0]concat=n='.count($playlist->audio).':v=0:a=1[out]\' -map \'[out]\' -vn -ac 2 -acodec aac -start_number 0 -hls_time 10 -hls_list_size 0 -f hls ';
+
+            $cmd .= $hlsDir . DIRECTORY_SEPARATOR .$playlist->type .'.m3u8';
+            $this->info($cmd);
+            $process = Process::fromShellCommandline($cmd);
+            $process->run();
+
+            if (!$process->isSuccessful()) {
+                throw new ProcessFailedException($process);
+            }
+
+            $playlist->status = Playlist::PLAYLIST_STATUS_COMPLETED;
+            $playlist->save();
         }
-
-        File::makeDirectory($hlsDir);
-
-        $cmd = 'ffmpeg ';
-        foreach ($playlist->audio as $audio){
-            $cmd .= ' -i ' . storage_path('app/'.$audio->path);
-        }
-        $cmd .= ' -filter_complex \'[0:0][1:0]concat=n='.count($playlist->audio).':v=0:a=1[out]\' -map \'[out]\' -vn -ac 2 -acodec aac -start_number 0 -hls_time 10 -hls_list_size 0 -f hls ';
-        $cmd .= $hlsDir . DIRECTORY_SEPARATOR .$playlist->type .'.m3u8';
-        $this->info($cmd);
-        $process = Process::fromShellCommandline($cmd);
-        $process->run();
-
-        if (!$process->isSuccessful()) {
-            throw new ProcessFailedException($process);
-        }
-
-        $playlist->status = Playlist::PLAYLIST_STATUS_COMPLETED;
-        $playlist->save();
     }
 }
